Extract Oracle persistence + migration helpers
- Move disk save/load + legacy candidate selection + wipe helpers into Oracle.Persistence.cs.
- Move migration/ensure helper methods (skills/research/presets/facilities) into Oracle.Migrations.cs.
- Keep Oracle.cs focused on runtime/gameplay state while preserving the same save pipeline + entrypoints.
- Update SaveSystemV11ImplementationLog with the new partial file map.
